requirements.txt
torch>=2.2
transformers>=4.42
datasets
accelerate
peft
bitsandbytes
jinja2
evaluate
rouge-score
sacrebleu
tensorboard
pyyaml
tqdm
scipy
numpy
pandas
scikit-learn
unsloth
tf-keras
sacremoses
python-docx

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Core runtime
requests>=2.32.0
tqdm>=4.66.0
PyYAML>=6.0.1

# Corpus processing
pdfminer.six>=20221105
python-docx>=1.1.2

# Tokenization (optional but recommended)
tiktoken>=0.7.0

# Validation & retrieval
sentence-transformers>=3.0.1
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.2.1
